Module Free.Make

The implementation of smart constructors.

Parameters

module P : Param

Signature

type dim = P.dim

The interval algebra.

type var = P.var

The type of cofibration variables.

type cof = ( dim, var ) Syntax.free

The type of freely generated cofibrations.

val var : var -> cof

Alias of Syntax.Free.var.

include Endo.S with type dim := dim and type cof := cof

val cof : ( dim, cof ) Syntax.endo -> cof

The embedding of cofibrations to cof.

val eq : dim -> dim -> cof

Smarter version of Syntax.Endo.eq that checks equality.

val eq0 : dim -> cof

eq0 r is eq r dim0.

val eq1 : dim -> cof

eq1 r is eq r dim1.

val bot : cof

The bottom cofibration.

val top : cof

The top cofibration.

val join : cof list -> cof

Smarter version of Syntax.Endo.join that simplifies cofibrations using syntactic criteria. For example, join [meet []] gives cof (Meet []).

Note that the simplification is attempting to strike a balance between optimality and efficiency, and thus it will not perform all possible syntactic reduction. To obtain more reduced cofibrations, use only smart constructors (instead of raw constructors) to build cofibrations.

val meet : cof list -> cof

Smarter version of Syntax.Endo.meet that simplifies cofibrations using syntactic criteria. See join.

val boundary : dim -> cof

boundary r gives a cofibration equivalent to join [eq0 r; eq1 r].

val forall : (dim * cof) -> cof

forall (r, cof) computes forall r. cof, using the syntactic quantifier elimination and potentially other simplification procedures used in eq, join, and meet.