Module Kado.Syntax

The abstract syntax of the restricted predicate logic of cofibrations.

type ('r, 'a) endo =

| Eq of 'r * 'r

| Join of 'a list

| Meet of 'a list

A family of polynomial endofunctors ('r, -) t indexed in an interpretation of the interval algebra 'r. This is for multiple types (for example, abstract syntax) to embed the langauge of cofibrations.

type ('r, 'v) free =

| Cof of ( 'r, ( 'r, 'v ) free ) endo

| Var of 'v

For each interval algebra 'r, we define the free monad ('r, -) t on the polynomial endofunctor ('r, -) endo: each ('r, 'v) t is the language of cofibrations over an interval algebra 'r, with indeterminates drawn from 'v.

module Endo : sig ... end

Stupid constructors for endo.

module Free : sig ... end

Stupid constructors for free.