Module Kado.Theory

The Theory module implements decision procedures for sequents relative to a theory over the interval, stated in the language of cofibrations.

module type Param = sig ... end

Parameters of cofibration solvers.

module type S = sig ... end

The signature of cofibration solvers.

module Make (Param : Param) : S with module Param := Param

The cofibration solver.