Module S.Alg

Algebraic theories over the interval.

type t = alg_thy

The type of an algebraic theory (no unreduced joins).

val empty : t

The empty theory.

val consistency : t -> [ `Consistent | `Inconsistent ]

consistency thy returns the consistency of the theory thy.

val split : t -> cof list -> t list

split thy cofs returns irreducible joins under additional cofibrations cofs.

val meet2 : t -> t -> t

meet2 thy1 thy2 computes the conjunction of the two theories thy1 and thy2. This is useful for supporting compilation units with top-level cofibration declarations.