Module S.Disj

Disjunctive theories over the interval.

type t = disj_thy

The type of a disjunctive theory.

val empty : t

The empty theory.

val consistency : t -> [ `Consistent | `Inconsistent ]

consistency thy returns the consistency of the theory thy.

val assume : t -> cof list -> t

assume thy cofs assumes cofs and returns the new theory.

val test_sequent : t -> cof list -> cof -> bool

test_sequent thy context cof tests the validity of a sequent against the supplied theory. Equivalent to assuming the cofibrations context and then testing truthness of cof.

val envelope_alg : Alg.t -> t

envelope_alg thy constructs the enveloping disjunctive theory of the algebraic theory thy.

val decompose : t -> Alg.t list

decompose thy returns irreducible joins as algebraic theories.

val simplify_cof : t -> cof -> cof

simplify_cof thy cof returns a potentially simplified cofibration that is equivalent to cof. This is useful for displaying sensible cofibrations to users. While it will remove useless cofibrations such as r=r and 0=1, it does not perform non-local simplification such as reducing x=1 /\ x=1 to x=1. Also, the simplification is an expensive process and should be used wisely.

val forall_cof : t -> (Param.dim * cof) -> cof

forall_cof thy (r, cof) computes forall r. cof with respect to the equations in the theory thy, using the syntactic quantifier elimination and potentially other simplification procedures used in Builder.Endo.S.eq, Builder.Endo.S.join, and Builder.Endo.S.meet. This is slower than Builder.Endo.S.forall which does not take equations into consideration.

Note: this is experimental and might be removed if it turns out that we do not need it. Also, r cannot be dim0 or dim1.

val meet2 : t -> t -> t

meet2 thy1 thy2 computes the conjunction of the two theories thy1 and thy2. This is useful for supporting compilation units with top-level cofibration declarations.