Module Free.Make

The implementation of smart constructors.

Parameters

module Param : Param

Signature

include Endo.S with type Param.cof := (Param.dim, Param.var) Syntax.free and module Param := Param

type t = Param.cof

The type of built cofibrations.

val cof : (Param.dim, t) Syntax.endo -> t

The embedding of cofibrations to t.

val le : Param.dim -> Param.dim -> t

Smarter version of Syntax.Endo.le.

val bot : t

The bottom cofibration.

val top : t

The top cofibration.

val join : t list -> t

Smarter version of Syntax.Endo.join that simplifies cofibrations using syntactic criteria. For example, join [meet []] gives cof (Meet []).

Note that the simplification is attempting to strike a balance between optimality and efficiency, and thus it will not perform all possible syntactic reduction. To obtain more reduced cofibrations, use only smart constructors (instead of raw constructors) to build cofibrations.

val meet : t list -> t

Smarter version of Syntax.Endo.meet that simplifies cofibrations using syntactic criteria. See join.

val eq : Param.dim -> Param.dim -> t

eq is equivalent to meet [le x y; le y x].

val eq0 : Param.dim -> t

eq0 r is equivalent to eq r dim0.

val eq1 : Param.dim -> t

eq1 r is equivalent to eq r dim1.

val boundary : Param.dim -> t

boundary r is equivalent to join [eq0 r; eq1 r].

val forall : (Param.dim * t) -> t

forall (r, cof) computes forall r. cof, using the syntactic quantifier elimination and potentially other simplification procedures used in le, join, and meet.

Note: r cannot be dim0 or dim1.

val var : Param.var -> t

Alias of Syntax.Free.var.