Module Syntax.Endo

Stupid constructors for endo.

type ('r, 'a) t = ('r, 'a) endo =

| Le of 'r * 'r

| Join of 'a list

| Meet of 'a list

A family of polynomial endofunctors ('r, -) t indexed in an interpretation of the interval algebra 'r. This is for multiple types (for example, abstract syntax) to embed the langauge of cofibrations.

val le : 'r -> 'r -> ('r, 'a) endo

le x y is Le (x, y)

val join : 'a list -> ('r, 'a) endo

join phis is Join phis

val meet : 'a list -> ('r, 'a) endo

meet phis is Meet phis

val bot : ('r, 'a) endo

bot is Join []

val top : ('r, 'a) endo

top is Meet []

val map : ('a -> 'b) -> ('r, 'a) endo -> ('r, 'b) endo

map is the functoriality of the endo functor ('r, -) t.

val dump : (Stdlib.Format.formatter -> 'r -> unit) -> (Stdlib.Format.formatter -> 'a -> unit) -> Stdlib.Format.formatter -> ('r, 'a) endo -> unit

Ugly printer.