Module Syntax.Free

Stupid constructors for free.

type nonrec ('r, 'a) endo = ('r, 'a) endo =

| Le of 'r * 'r

| Join of 'a list

| Meet of 'a list

A family of polynomial endofunctors ('r, -) t indexed in an interpretation of the interval algebra 'r. This is for multiple types (for example, abstract syntax) to embed the langauge of cofibrations.

type ('r, 'v) t = ('r, 'v) free =

| Cof of ('r, ('r, 'v) free) endo

| Var of 'v

For each interval algebra 'r, we define the free monad ('r, -) t on the polynomial endofunctor ('r, -) endo: each ('r, 'v) t is the language of cofibrations over an interval algebra 'r, with indeterminates drawn from 'v.

val var : 'v -> ('r, 'v) free

var v is Var v

val cof : ('r, ('r, 'v) free) endo -> ('r, 'v) free

cof phi is Cof phi

val le : 'r -> 'r -> ('r, 'v) free

le x y is Cof (Le (x, y))

val join : ('r, 'v) free list -> ('r, 'v) free

join phis is Cof (Join phis)

val meet : ('r, 'v) free list -> ('r, 'v) free

meet phis is Cof (Meet phis)

val bot : ('r, 'v) free

bot is Cof (Join [])

val top : ('r, 'v) free

top is Cof (Meet [])

val dump : (Stdlib.Format.formatter -> 'r -> unit) -> (Stdlib.Format.formatter -> 'v -> unit) -> Stdlib.Format.formatter -> ('r, 'v) free -> unit

Ugly printer.